back_end/requirements.txt
asgiref==3.6.0
backcall==0.2.0
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
django-import-export==3.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
executing==1.2.0
importlib-metadata==6.1.0
install==1.3.5
joblib==1.2.0
Markdown==3.4.3
MarkupPy==1.14
nest-asyncio==1.5.6
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
tablib==3.3.0
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
wcwidth==0.2.6
xgboost==1.7.4
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
